1번 회원 관리자로 지정, 게시물 수정/삭제 권한체크

+++ src/main/java/com/sbs/springPractice/controller/UsrArticleController.java
@@ -78,7 +78,13 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		
+		if ( loginedMemberId == 0 ) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}		
+		
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
@@ -88,28 +94,48 @@ public class UsrArticleController {
 		if (article == null) {
 			return new ResultData("F-1", String.format("%d번 게시물은 존재하지않습니다.", id));
 		}
+		
+		ResultData getActorCanDeleteRd = articleService.getActorCanDeleteRd(article, loginedMemberId);
+		
+		if ( getActorCanDeleteRd.isFail()) {
+			return getActorCanDeleteRd;
+		}
+
 
 		return articleService.deleteArticle(id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
-
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		
+		if ( loginedMemberId == 0 ) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+		
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
-		if (title == null) {
-			return new ResultData("F-1", "title을 입력해주세요.");
-		}
-		if (body == null) {
-			return new ResultData("F-1", "body를 입력해주세요.");
-		}
-
+		
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return new ResultData("F-1", String.format("%d번 게시물은 존재하지않습니다.", id));
+		}
+		
+		if (title == null) {
+			title = article.getTitle();
+		}
+		
+		if (body == null) {
+			body = article.getBody();
+		}
+		
+		ResultData getActorCanModifyRd = articleService.getActorCanModifyRd(article, loginedMemberId);
+		
+		if ( getActorCanModifyRd.isFail()) {
+			return getActorCanModifyRd;
 		}
 
 		return articleService.modifyArticle(id, title, body);
